requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
aiosqlite>=0.20.0
httpx>=0.28.0
python-dotenv>=1.0.0
